+++ src/triangulation/mod.rs
@@ -1,3 +1,5 @@
 mod delaunay2;
+mod ids;
 
 pub use delaunay2::{Triangulation2, delaunay2};
+pub use ids::{EdgeId, FaceId, VertexId};